Valid

+++ src/cli/Manager/Controller/ControllerDelete.php
@@ -4,18 +4,30 @@ namespace Mora\Core\cli\Manager\Controller;
 use Mora\Core\cli\Manager\Routes\CustomRoutesManager;
 use Mora\Core\Config\JsonConfigManager;
 use Mora\Core\cli\Manager\Firewall\FirewallManager;
+use Mora\Core\cli\Helpers\Methods;
 
 class ControllerDelete{
     public static function delete($name){
         $path = CONTROLLER . "/$name" . "Controller.php";
         if(file_exists($path)){
-            unlink($path);
-            self::deleteRoute($name);
-            self::deleteFirewallTarget($name);
+            if(self::confirmDelete()){
+                unlink($path);
+                self::deleteRoute($name);
+                self::deleteFirewallTarget($name);
+                ControllerMessage::delete_success($name);
+            }
+        }else {
+            ControllerMessage::controller_not_found($name);
         }
     }
+    public static function Interactive(){
+        $name = Methods::ask("ask_controller_name");
+        $name = ucfirst(strtolower($name));
+        self::delete($name);
+    }
     public static function confirmDelete(){
-        
+        $answer = Methods::ask("controller_confirm_delete");
+        return ($answer == "" || $answer == "yes");
     }
     private static function deleteRoute($name){
         $path = CustomRoutesManager::$path;
